[display][fidl] VSync events must have valid DriverConfigStamps.
This CL updates the FIDL interface comments and the Coordinator to
reflect the API change in the attached bug.

The fake-display driver used in unit tests has already been updated to
reflect the change, so all unit tests pass.

The virtio-gpu-display driver referenced in manual tests has not been
updated to filter VSync events, so the manual tests prove that the
end-to-end system still works with drivers that haven't been updated.
This is not a surprising result, as VSync events for a display are only
forwarded to Coordinator clients after they apply a configuration to
that display.

Bug: 401388764
Test: fx test //src/graphics/display
Test: Booted workbench_eng.x64 with ffx emu start --engine qemu
Test: ffx target ssh display-tool squares
Change-Id: Ia5ad38837c1c8f5603095c3927b45bc96234b0e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1329640
Reviewed-by: Yilong Li <liyl@google.com>
API-Review: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>