[starnix][nanohub] Add FIDL-backed display sysfs file
This CL introduces a new FIDL-backed sysfs file for interacting with
the display service. The primary motivation for this is to allow the
display to be reset from the command line.

This CL also refactors the existing sysfs file handling to be more
generic. The previous implementation was specific to nanohub, but the
new implementation can be used for any sysfs file.

Bug: b/420740621
Test: fx test starnix_nanohub_module_test
Change-Id: I65e5709a6f34723966bfe0a9c8c0fb3a4fe212d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1327287
Reviewed-by: Eric Rahm <erahm@google.com>
Reviewed-by: Chad Norvell <chadnorvell@google.com>
Commit-Queue: Lingxue Luo <lingxueluo@google.com>