[display][coordinator] Remove SendVsyncsAfterClientsBail test case.
In the SendVsyncsAfterClientsBail test case, the display engine driver
sends a `OnDisplayVsync` event with an invalid config stamp, and the
coordinator is expected to ignore that that event.

This is not allowed anymore in the latest EngineListener protocol.
Thus, this change deletes this test case.

Test: display-coordinator-unittests
Bug: 336394440
Change-Id: Ibf4c2919c12473ddc4c4aeb6483c098305bff152
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1331044
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Victor Costan <costan@google.com>