[kernel][sched] Prevent time skew in accounting and tracing
This commit fixes two sources of time skew in scheduler accounting and
trace events.

Accounting Skew from Lock Contention

When the monotonic clock is sampled *before* acquiring a run-queue lock,
high contention can make the timestamp stale. This can cause incorrect
calculations.

For example, a CPU could compute a negative runtime if:

1. CPU A samples time `T1` before rescheduling.
2. CPU B samples time `T2` (where `T2 > T1`) before a PI operation on
   CPU A's active thread, acquires the lock first, and updates CPU A's
   `last_started_running` to `T2`.
3. CPU A finally acquires the lock and calculates runtime using its old
   timestamp: `T1 - last_started_running` (now `T1 - T2`), resulting in
   a negative value.

This is fixed by moving the time sample to *after* the run-queue lock is
acquired.

Trace Event Skew

A secondary skew occurs when trace counters (e.g., energy usage) are
calculated using monotonic time deltas, while the corresponding trace
events are timestamped using the boot clock. This mismatch reduces the
accuracy of offline analysis tools (not considering the deviations due
system suspend, which are an independent consideration).

This is fixed by sampling the monotonic and boot clocks together and
using the coherent pair for both the calculation and the event
timestamp.

For convenience, adding the skew-sensitive trace counter has been pulled
into this patch, since it is standalone from the larger patch is was in,
and is more closely related to this overall fix.

Bug: b/379576294
Bug: b/323269528
Change-Id: If7266a8dc2d92818acd0faf81cae4f1210bc3d40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1328905
Commit-Queue: Corey Tabaka <eieio@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>