[display][bin] Fix display-tweak packaging.
This CL removes a package definition that didn't include the .cml, and
fixes the .cml that wasn't validated before.

The manual tests verify that the new package has the capabilities that
it needs correctly routed to it in the shell environment.

Bug: 388885807
Test: Booted internal product with backlight support
Test: ffx target ssh -- display-tweak backlight --power on --brightness 1.0
Test: ffx target ssh -- display-tweak backlight --power on --brightness 0.25
Change-Id: I0c6e62e00e87279faa64289fcfb2a2dff31076f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1331412
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>