[fastboot] reduce userspace fastboot buffer size
Previously userspace fastboot would allow download buffer allocations up
to the entire size of system memory, which only works if the device has
enough RAM to fully contain any image we might flash to it.

For low-memory devices this is not the case, and attempting to allocate
too much memory gums up the system and either hangs or crashes. This CL
reduces the max allocation to 2% of the total RAM, which is going to be
in the O(10MiB) range for low-memory devices and O(100MiB) for
high-memory.

2% is fairly conservative I think, we may be able to bump it up once
we figure out what image userspace fastboot is running on and get a
better sense of what resource usage looks like.

Bug: b/426545225
Change-Id: I91d6f0638d297a54106be2f3f00ebe0b1b02a7b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1331407
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>