[kernel][port] Fix race condition in zx_port_cancel_key
This closes a race condition where zx_port_cancel_key could report that
it successfully canceled a packet that was in the process of being
returned to zx_port_wait on another thread. The root cause of this race
is that when an observer match triggers a packet is queued in two
operations:

1. PortDispatcher::Queue acquires the dispatcher's lock and adds the
   packet to the packet queue, then drops the lock
2. PortDispatcher::MaybeReap acquires the dispatcher's lock and then
   removes the observer from the dispatcher's list of observers.

This means that there is a window of time where an observer is both in
the port dispatcher's packets_ queue and observers_ list. The
PortDispatcher::CancelKey function was (incorrectly) assuming that if it
detected a matching observer in the dispatcher's observers_ list that
the observer was still pending.

This closes the window by merging queuing a packet and removing an
observer from the port dispatcher's observer list into one operation
that happens with the port dispatcher's lock continuously held. This
establishes the invariant that CancelKey was expecting that only live
observers are kept in the observers_ list.

This also splits the Queue calls for manually queued packets from those
triggered by PortObserver triggered waits. Callers that want to queue a
packet directly call PortDispatcher::Queue and no longer have to specify
a signals parameter.

Fixes: 427550165
Change-Id: I019f3cf11878d555d85aa1cc9423725b013c1441
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1308018
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: James Robinson <jamesr@google.com>