[trace-reader] Replace fbl::HashTable with std::unordered_map
This change continues the effort to port the trace-reader library
from FBL to the C++ standard library by replacing fbl::HashTable with
std::unordered_map.

This removes the need for the ProviderInfo struct to inherit from
fbl::SinglyLinkedListable and removes the fbl dependency from the
library and its tests. A direct dependency on //sdk/lib/fit is added
to replace the transitive one that was removed.

Bug: 432527736
Test: fx test host_x64/trace-reader-test-test
Change-Id: I8a05b151b09929e7c4fc8e63192f9cdc8dfab6bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1333105
Commit-Queue: Chris Masone <cmasone@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>