[network] Move TcpInfo out of fuchsia.posix.socket
...into a new fuchsia.net.tcp FIDL library.

We're going to use this struct in another FIDL API, so having it in
fuchsia.posix.socket no longer makes sense.

This change is safe, despite being API-breaking, because
fuchsia.posix.socket is not exposed in the public SDK. The only OOT
usage is in the Go runtime, but this change shouldn't be ABI-breaking,
so we don't need to worry about synchronizing the two changes.

Bug: 433947439
Change-Id: I6a6a69642e16c5af4b9d031b94010765081e4352
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1332584
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>