[display][coordinator] Implement EngineDriverClientFidl methods.
This change implements the `EngineDriverClientFidl`, which is
responsible for communicating with display engine drivers over the
`fuchsia.hardware.display.engine` FIDL protocol.

The implementation covers most methods of the `EngineDriverClient`
interface except for CheckConfiguration, ApplyConfiguration and
CompleteCoordinatorConnection.

This change is created using gemini-cli and manually reviewed.

--PROMPT--

Now work on EngineDriverClientFidl. First implement the
functions that are not supported yet. Skip CheckConfiguration and
ApplyConfiguration, CompleteCoordinatorConnection and UnsetListener.

If FIDL transport fails, for functions that has void return type, log it
but do NOT crash. For functions with return values, log it and then
crash using ZX_ASSERT.

Now implement a test for this class.

Step 1: Write a MockFidlEngine similar to
@mock-fidl-display-engine-listener.h and
@mock-fidl-display-engine-listener.cc , and put it in testing directory.

Step 2: Implement engine-driver-fidl-client-test.cc for the implemented
methods.

Now you are working on EngineDriverClientFidl test in
@engine-driver-client-fidl-test.cc . This test tests the class defined
in @engine-driver-client-fidl.h .

Currently we test each function that overrides the EngineDriverClient
interface.

For the functions that return zx::result<...>, we only test the case
where it returns a successful value, but we don't test the case where it
returns an error.

We need to add test cases for error handling.

For example, for function `zx::result<> ReleaseCapture(...);`, We
currently have test `TEST_F(EngineDriverClientFidlTest, ReleaseCapture)`
which tests the case where the engine driver replies a `success` result.

We need to:
1) Rename the test to `TEST_F(EngineDriverClientFidlTest,
   ReleaseCaptureSuccess)`
2) Create a test `TEST_F(EngineDriverClientFidlTest,
   ReleaseCaptureFailure)` which tests the case when the completer
   replies an error (but doesn't close the channel). We should make sure
   that the error received by the client is the same as the error sent
   by the engine driver.

Do this for each test, one at a time. Write down the expected new test.

DO NOT DELETE CODE. If you really have to delete code, write down the
rationale and think twice.

Before you modify the code, (1) Write an example of what you need to do
with ReleaseCapture(). (2) List all the tests that you need to modify.

--PROMPT--

Bug: 314126494
Test: fx test //src/graphics/display
Change-Id: If4484208cc527bea06858b17657fa9760eb7f237
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1331513
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>