[display][coordinator] Add tests for EngineDriverClientBanjo.
This change introduces unit tests for the `EngineDriverClientBanjo`,
which is the Banjo-based implementation for communicating with the
display engine driver.

To facilitate testing, a `MockBanjoEngine` is added. This mock simulates
the `fuchsia.hardware.display.controller/DisplayEngine` Banjo protocol,
allowing for isolated testing of the client logic.

The new tests cover all public methods of `EngineDriverClientBanjo`,
ensuring that Banjo calls are correctly translated and forwarded.

This change is created using gemini-cli and reviewed manually.
--PROMPT--
You are working on the @src/graphics/display/drivers/coordinator driver.

Follow examples in the following files mentioned, implement tests for
EngineDriverClientBanjo (defined in @engine-driver-banjo-client.h) in
@engine-driver-banjo-client-test.cc .

Step 1: Write a MockBanjoEngine similar to
@mock-banjo-display-engine-listener.h and
@mock-banjo-display-engine-listener.cc , and put it in `testing`
directory. DO NOT USE gMock. Implement the mock methods in the same way
as mock-banjo-display-engine-listener.h/cc.

Step 2: Implement engine-driver-banjo-client-test.cc for the implemented
methods of `EngineDriverClientBanjo`.

The list of test cases should be similar to
@engine-driver-fidl-client-test.cc .

Please follow the examples in @display-engine-events-banjo-test.cc to
implement each test case.
--PROMPT--

Bug: 314126494
Test: fx test //src/graphics/display

Change-Id: I5f3978422519c6b55f82b5866f72be68444b34d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1331514
Commit-Queue: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>