[fxtest] Support quiet mode for program events.
This CL expands the program started event to tag it with quiet_mode if
we want to signal to viewers that the program should not appear in
status output. This is useful for background processes that do not need
to be shown to the user persistently.

Bug: 435051200
Change-Id: I1f3bde5f340d91041ccfda350415dcdcc693587f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1334184
Reviewed-by: Dale Sather <dalesat@google.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>