[ffx] refactor doctor_summary()
For some upcoming ffx-doctor changes, it will be helpful to not
have a single 700-line function contain the bulk of the logic.

In breaking the function up, it became clear that there was an issue
where the function was returning too early, resulting in the
overall summary being skipped. So the tests needed to be updated a
bit as well.
 
Change-Id: I6a6a69641bbb74f9a6786190441594c2d16d9fb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1332961
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>