[windowed-stats] Avoid panic on drain
Currently, when we attempt to drain the calls for a time series, we
panic if the name of the time series is not present in the map. For
testing, it is helpful to assert that the time series that we expect to
have data, do have data, but also to assert that the time series that we
do not expect to have data, do not have data.

Instead of panicking when the name is not present, just return an empty
vector to allow callers to assert that the time series was not involved
in any calls.

Change-Id: Ib9abff221c953f34887b8b472f653d30fe0c49f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1333626
Reviewed-by: Sean Olson <seanolson@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>