[tests] Change the default create-no-exception-channel
I thought this was done in https://fxrev.dev/1331424 but that didn't
account for the fact that the test_spec option is actually coming
through tests.json, not test-list.json.

When tests.json doesn't specify it the default wasn't overridden in the
prior CL. This actually changes the default behavior, and when the
flakes are resolved we can revert this change.

Bug: 435166699
Bug: 434722978
Change-Id: I6a6a69644e8c33fe716bed7f605a0721d893dd29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1334624
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>