[reachability] Add structure to report time series
Create a processors module in reachability's telemetry library to set up
the structure to report LinkProperties and LinkState based on interface
identifiers.

The first InterfaceTimeSeriesGrouping is the InterfaceType.
InterfaceTypes must be provided upon creation of the LinkPropertiesState
logger, and time matrices will be created as part of the instantiation
of the logger. Future additions to InterfaceTimeSeriesGrouping (such as
NICID/Name) will be dynamic, as in, the time matrices will be created as
new NICIDs and interface names are observed.

Based on the interface identifiers of the
update, the processor will report LinkProperties or LinkState for
matching interface identifier entries in the Map.

Add tests to ensure that the inspect nodes are working as expected, and
tested that nodes are present locally.

See https://fxbug.dev/432298588 for more details.

Bug: 432298588
Fixed: 432298086
Change-Id: Ic71a809c128cd2b3d11b3be03bafa49cf54a84c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1307170
Reviewed-by: Kiet Tran <kiettran@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>