[audio][adr] Accounting fixes in RingBuffer memory
When the audio driver method `RingBuffer::GetVmo` is called,
a driver is allowed to allocate a larger buffer than the
`min_frames` passed to it by the client. Previously when
this occurred, the buffer descriptor contained values for
`consumer_bytes` and `producer_bytes` that reflected the
requested sizes, rather than the actual sizes.

With this CL, any overallocation by the driver is explicitly
provided to the client. (If the driver needed a larger area
within the ring buffer, it should have specified a larger
value for `driver_transfer_bytes`.)

Thus, following this change it is now invariant for all
`fuchsia::audio::RingBuffer`s returned by calls to
`fuchsia_audio_device::Control::CreateRingBuffer()` that
their `.buffer.size` exactly equals `.producer_bytes` plus
`.consumer_bytes`.

This makes it possible for producers and consumers to safely
operate in "high-water" mode where the entire ring buffer is
used, to either maximize glitch-resilience (in traditional
"top-up" high-water mode) or minimize system/power load (in
"long-sleep" high-to-low-water mode).

Bug: 358640647
Bug: 358430098
Bug: 425376354
Bug: 432252597

Change-Id: I7b00652f64e7a91987d988c2de563c420ff608d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1329159
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>