[async-loop] Switch to zx_port_cancel_key
This updates the async-loop library to use zx_port_cancel_key() to
cancel pending waits.

This means that the async_wait structure's `object` field is no longer
used by the async loop library. Many users of this library use this
field to store data for other uses, so the field is left alone. We may
consider removing this field in the future to reduce the size of this
structure in memory.

This is a reland of https://fxrev.dev/1046693

Change-Id: If3a673f2523f58f89a46ee7bce0de5b3b9a56752
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1300183
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Mega-CQ: James Robinson <jamesr@google.com>