[starnix] Stub IORING_SETUP_COOP_TASKRUN for io_uring
This change implements the  flag for the
 syscall. This flag is currently a no-op, but it is
required by some applications that use io_uring.

The implementation adds the flag to the list of ignored flags, so that
it doesn't return an error when used.

This is a reland of
https://fuchsia-review.googlesource.com/c/fuchsia/+/1334364 after
io_uring syscalls had been added for aarch32.

Bug: 297431387
Change-Id: Icb6c9cc03399aed2f4f09156d8ca577bdbdb585a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1334786
Commit-Queue: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>