[starnix] Support UFFD_FEATURE_SIGBUS and basic userfault ioctls
Support the minimum set of userfaultfd features needed for the garbage
collection. This includes:
 * UFFD_FEATURE_SIGBUS, which sends SIGBUS when a uffd-registered page
   was accessed without being populated
 * UFFDIO_COPY and UFFDIO_ZEROPAGE that can be used to populate the
   uffd-registered range.

For now, implements these features by altering mapping protections and
forcing the page faults. This approach is not very performant and
suffers from data race issues, but it should suffice as a first
approximation. It is not used in any production code for now and is only
exercised in tests.

Test: userfaultfd_test.cc
Bug: 297375964, 348485195

Change-Id: I88d14234524536878b7535c7b2cdb109cd96a04e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213925
Commit-Queue: Maria Glukhova <mariagl@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>