[netstack3] Propagate PMTU errors even if in cache
Currently Netstack3 only propagates PMTU errors (PacketTooBig or
FragmentationRequired) to the transport layer if the PMTU error resulted
in a new PMTU being inserted in the PMTU cache. This means that once the
cache has been updated, subsequent traffic that triggers the error does
not receive it (such as a TCP socket), which means it can't respond
correctly (in the case of TCP, by updating its MSS).

Fix this by always returning the current PMTU from
PmtuHandler::update_pmtu_if_less and PmtuHandler::update_pmtu_next_lower
so that even if the error didn't result in an update to the cache, the
current PMTU will be propagated to the original sender that caused the
error.

Bug: 422307328
Fixed: 435260334
Test: netstack3-core-ip-test
      netstack3-core-ip-integration-test
      netstack-socket-integration-test
Change-Id: Ic8f52640cee32907c7d97ee2c8500b5942deb954
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335018
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>