[fuchsia][macro] fix macro hygiene issues
No longer impl quote::ToTokens for Executor

When we were tokenizing the Executor we were shadowing the `func`
variable which was causing hygiene issues when invoking the proc macro
from within a decl macro.

Now have a `build_token_stream` function that takes a reference to a
syn::Ident identifier, and returns a TokenStream.

The syn::Ident is is a well-known syn::Ident built in the
Transformer::finish() function so it can be safely used through the code
generation.

Bug: b/42157685
Change-Id: I19f908488e343497c484e629c192e37c6a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335704
Reviewed-by: Marie Janssen <jamuraa@google.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>