[camera] Remove _LIBCPP_ENABLE_HARDENED_MODE
This setting has been a no-op to enable a hardening mode and a change
in upstream libc++ now detects and warns when this macro is set. This
CL removes the macro until a hardening mode strategy is set for Fuchsia.

Bug: 42082322, 435515222
Change-Id: I08d2147df58b8e5e2fc0c37b61f87ac2bc81a9fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335990
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: John Sasinowski <jsasinowski@google.com>