[zircon] Remove _LIBCPP_ENABLE_HARDENED_MODE
This setting has been a no-op to enable a hardening mode and a change
in upstream libc++ now detects and warns when this macro is set. This
CL removes the macro until a hardening mode strategy is set for Fuchsia.

Bug: 42082322
Fixed: 435515222
Change-Id: I548ee06ea48309236ad7c3b94bdac3721bd1c100
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335991
Reviewed-by: Dustin Green <dustingreen@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>