[kernel][vm] change the PAGE_SIZE define to unsigned
For historical reasons lost in history, the kernel's PAGE_SIZE was
defined as signed, which really isn't accurate. Change the
definition of PAGE_SIZE and all of the derivatives of it to be
unsigned long (64bit unsigned) which makes all of the math, masks,
and inverted masks generated from it proper.

Change for all three architectures.

Change-Id: Ic25bf76584d37e977538550010e7631c7221de80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335008
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>