[ffx] fix doctor when query matches >1 target
The code to invoke `ffx target show` was attempting to specify the
target by updating the command line (`ffx.global.target`), but that
doesn't work -- it is actually pulled in from config. Since we
actually want to change the target that each invocation uses,
the simplest approach was to add a new mechanism for overriding
the target in specific copy of an EnvironmentContext.

Fixed: 434772377

Change-Id: I6a6a6964b374e8525d5ede6b7f1db8bc495f037c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335796
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>