[starnix][mm] Factor AddressSpace into a new file
This CL factors a bunch of code out of memory_manager.rs related to the
state of the address space into an address_space.rs. This code will
become more complicated as we change the representation of the
RangeMap to improve memory efficiency.

Ideally, the RangeMap would be an implementation detail of the
AddressSpace struct, but further work will be needed to make that
happen.

Change-Id: I30a16b116abadf7432bdce57adbca93469e69a01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1336696
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>