[build][idk] Use IDK manifest for clidoc
Previously, the clidoc build step used an internal IDK build manifest to
identify the tools to process. Instead, use the manifest included in the
IDK since the build manifest is being removed. The output will also
better match the contents of the IDK containing the documentation.

As a result, `blobfs.md` is renamed to `blobfs_do_not_depend.md`,
matching the name of the binary in the IDK.

Bug: 407083737
Test: The contents of `obj/tools/docsgen/clidoc_out` are the same before and after the change except for the `blobfs` name change mentioned above.
Change-Id: Ife10082c96c77d1e110a688a73b49d3702eb133f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1333838
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>