[ffx] Fix Mutex::lock() across await
There was some code that held a lock across an await point. It
was harmless, since the lock was a local variable, so there never
would be any contention. The fix was to change it from a Mutex to
a RefCell.

At the time this bug was filed, there were a few other potential
lock/await problems, but they were all fixed by other CLs.

Fixed: 432297777

Change-Id: I6a6a6964bad17d0a4eb4cb6d3e3bf5443511323b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1337944
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>