[display][amlogic] Use ModeId to determine display mode.
This change updates the amlogic-display driver to use the `mode_id`
from the display configuration to determine the target display mode.

The `CheckConfiguration` and `ApplyConfiguration` functions are
refactored to inspect the `mode_id`. If a valid `mode_id` is present,
the driver uses the corresponding `display::Mode` for validation and
applying the configuration.

Only if no `mode_id` is provided, the driver falls back to the legacy
behavior of using the `display_timing` struct. This also involves
refactoring internal functions to operate on the `display::Mode` C++
type instead of raw Banjo structs.

Test: fx test //src/graphics/display
Test: On workbench_eng.vim3 (with circular display), "display-tool
      squares" worked.
Test: On core.nelson, virtcon booted. (amlogic-display locally modified
      to perform full hardware reset on boot)
Bug: 435313315
Change-Id: I282627ef40527023edeb716336432b3a8cae61e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335616
Reviewed-by: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>