[ffx][trace] Recording requested categories when tracing
This keeps track of the categories requested on the commandline
so the resulting trace can be inspected, and warnings generated if
requested categories do not appear in the trace data.

This was done slightly differently in the daemon only implementation,
where the traceConfig object contains the requested categories and the
ffx daemon would expand them. Since we are moving away from the daemon
implementation, the categories are expanded by the trace plugin and
passed in the TraceConfig object, and the requested (un expanded)
categories are passed in the TraceOptions.

The requested categories are held in the TraceTask object, so they
can be returned on `trace stop`.

Bug: 433284236
Change-Id: I90eb6e6b129b6ba8112bd45535e1e4f55f5feb29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1327616
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
API-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>