[sdmmc][sdhci] Log buffer count for data timeout errors
Data timeout errors are occasionally seen in CI/CQ. Log the buffer count
when this happens so that we can tell if the reads that fail are packed
commands. If so, more debugging information can be added to try to
narrow down the issue.

Bug: 427293779
Test: CQ
Change-Id: Ia6755d838c839afbc8dd7d505dc66be1d974f91c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1337725
Fuchsia-Auto-Submit: Braden Kell <bradenkell@google.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>