[microbenchmarks][starnix] Switch from host toolchain to linux
This uses the linux toolchains to build the version of the
microbenchmark binaries run in Starnix instead of relying on the host
toolchain. This is a useful cleanup on its own and also will make it
easier to set up aarch32 builds.

Bug: 297527377
Change-Id: I954c65bb6e414638ebf8ff6bdc2fdc98f77a76b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1335011
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: James Robinson <jamesr@google.com>