[starnix][mm] Simplify mlock address calculations
Previously, when we locked memory in a mapping that did not include the
beginning of a mapping, we mutated the memory_offset field in a copy of
the mapping and then used that value to find the offset in the backing
VMO. This approach created a Mapping object in an inconsistent state.

Instead, we now use the address_to_offset method of Mapping to do the
calculation directly.

Bug: b/435494969
Change-Id: I776201570c3a3f6986cfa60fad62a4b955f4eaa2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1337804
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Adam Barth <abarth@google.com>