[display][api-protocols] Support both ModeId and DisplayTiming.
This change refactors the `DisplayEngineInterface` to support both
predefined display mode IDs and custom display timings.

The `CheckConfiguration` and `ApplyConfiguration` methods are updated
to accept a `std::variant<ModeId, DisplayTiming>`. This allows drivers
to handle display configurations specified by either a `ModeId` or a
`DisplayTiming` struct, with `ModeId` being the preferred method if
available. This is critical for supporting display engines that do not
have predefined modes but use EDID to get timings.

The Banjo and FIDL adapters are updated to populate this variant,
allowing in-tree drivers to get either a `ModeId` or a
`DisplayTiming` when handling configs.

Test: fx test //src/graphics/display
Change-Id: I60d7b868f421c3d03ce8bbd3257a5e3110e904fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1336846
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>