[pci] Remove LEGACY_NOACK interrupt mode
With the implementation of "RFC-0270: Zircon Virtual Interrupt Signals"
there is no longer a need for a PCI protocol specific ack, as the bus
driver is now able to handle this using the virtual interrupts handed
out to driver clients. This will reduce the complexity of handling
interrupt dispatch when the RFC functionality is implemented in the
driver.

It was also unused, so it's nice to delete it.

Fixed: 421222319

Change-Id: I411ca261b61ca681d453d58f3ca82e880cfed89e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1288496
API-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>