[ffx] implement ffx --direct doctor
When in direct mode, "ffx doctor" still will check the daemon,
but it is not an error for the daemon to be unavailable. The main
difference is that all target validation is done through direct
connections rather than through the daemon. The implementation is
straightforward: it simply uses the diagnostics already implemented
as "additional diagnostics" -- but we use them as the "primary"
target check instead. Other differences in direct mode:

* we don't do a "compatibility" check, because that is an Overnet
  concept
* we always check connectivity via FDomain

Change-Id: I6a6a696486de41e7719e08201341087d2315e5bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1336905
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>