[starnix][mm] Move mremap off of memory_offset()
This code is better off using `address_to_offset`. This CL is part of a
series that removes all the clients of `memory_offset` so that we can
change the internal representation of MappingBackingMemory to use less
memory.

Bug: b/435494969
Change-Id: I7be716fe7861a918be1d9827be9da6d9e3bff0e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1337951
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Adam Barth <abarth@google.com>