[fxfs] Tombstone closed files during overwrite.
The files were added to the graveyard and all the complex swapping of
vmo ownership were handled, but the boring path of overwriting a blob
that was not open at all had been forgotten. Added more tests so that
all three code paths (blob has a vmo opened, the blob is open but no vmo
is shared, and the blob is closed) now verify cleanup and usability of
the new blob.

Bug: 430394762
Change-Id: I91fb58e451b6b98e12684188f2b5fd6489bbd24f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1333484
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>